Add files via upload
Build a Celestial Bodies Database
This is one of the required projects to earn your certification. For this project, you will build a database of celestial bodies using PostgreSQL.

This course runs in a virtual Linux machine using Gitpod. Follow these instructions to start the course:

Create a GitHub account if you don't have one
Click the start button below
Login to Gitpod with your GitHub account if you aren't already
Once the virtual Linux machine is finished loading, start the CodeRoad extension by:
Clicking the "hamburger" menu near the top left of the VSCode window,
Going to the "View" menu,
Clicking on the "Command Palette" option,
and running the "CodeRoad: Start" command
Follow the instructions in CodeRoad to complete the course
Complete both steps below to finish the challenge.
Step 1: Complete the project
The project runs in a virtual machine, complete the user stories described in there and get all the tests to pass to finish step 1.
Important: After you pass all the project tests, save a dump of your database into a universe.sql file so you can complete step 2. There will be instructions how to do that within the virtual machine.